requirements.txt
glfw
PyOpenGL
Pillow